Added accept and reject commands
Closes #1

These commands are focused on project managery stuff.

`fbz accept` is used to close a resolved case, reflecting that
the work performed for that case is acceptable (the related
PR is merged, testing has shown good results, etc).

`fbz reject` is used to reactivate a resolved case, reflecting
that it needs more work before it can be accepted.

Both of these commands require a commment, but most particularly,
it would be really uncool to `reject` without an explanation.